fix(compression): cast to avoid warning

+++ includes/acl/compression/impl/compress.impl.h
@@ -556,7 +556,7 @@ namespace acl
 				ACL_ASSERT(written_bulk_data_size == bulk_data_size, "Unexpected amount of data written"); (void)written_bulk_data_size;
 				db_header->bulk_data_hash = hash32(db_header->get_bulk_data(), bulk_data_size);
 
-				ACL_ASSERT((database_buffer - database_buffer_start) == database_buffer_size, "Unexpected amount of data written");
+				ACL_ASSERT(uint32_t(database_buffer - database_buffer_start) == database_buffer_size, "Unexpected amount of data written");
 
 #if defined(ACL_HAS_ASSERT_CHECKS)
 				// Make sure nobody overwrote our padding
